A Cookie implementation to simulate a logged user.
In this commit the navigation bar will show some references according to
a user login status. By the moment no username or email are required to
loggin the user. At this commit it is just necessary to psuh the login
button in the loggin view.

The cookie is read in every request.
The login view was built and some css files were updated to be adapted
to the new project update. Finally Here a regex is used to avoid find
the cookie value from an array of cookies which was transformed from a
string and then extracting the value from the found entry by splitting,
etc... I my opinion a regex was a better option to extract the cookie
this time. It is necessary to mention that the implemented regex pattern
is not optimal an can be improved and adapted according to the use case.