Now the web app can register and login users.
The hardcode that created a new user in case it did not exist was
removed and now the user creation is handled by the auth controller.
The user credentials are now stored in the database taking into account
to encrypt the password before save it.
Users can create its account thanks to the new view Signup. Validation
and routes protection are not implemented by the moment but these are
characteristics that will come in future commits.
Finally the css and the .ejs view files were updated to behave very well
in mobile and desktop environments.