Reset Password feature added!
Using the "crypto" native module a random token was generated.
This token was used to generate a secret link where the user could
reset/update its account password.
The user model was updated to handle the resetToken and its expiration.
Given that by the moment a emailing feature has not being implemented
the secret link is displayed at the console.
the resetToken and expirationDate are verified before let the user
update its password. If the token is already expired, then the secret
link is not reacheable and the user is redirected to the reset password
view to generate a new link.
After the user update its password the resetToken and its expiration
date are purged from the user data.